// workspaces/ballerina/ballerina-visualizer/src/components/ConnectorSelector/types.ts
/**
 * Copyright (c) 2025, WSO2 LLC. (https://www.wso2.com) All Rights Reserved.
 *
 * WSO2 LLC. licenses this file to you under the Apache License,
 * Version 2.0 (the "License"); you may not use this file except
 * in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

import { FlowNode, SearchKind } from "@wso2/ballerina-core";

export type ConnectorType = 'MODEL_PROVIDER';

export interface ConnectorTypeConfig {
    fieldLabel: string;
    fieldDocumentation: string;
    valueTypeConstraint: string;
    createButtonLabel: string;
}

export interface ConnectorInfo {
    text: string;
    description?: string;
    codeCommand?: string;
}

export interface ConnectorSpecialConfig {
    infoMessage?: ConnectorInfo;
    shouldShowInfo?: (symbol: string) => boolean;
}

export interface BaseConnectorProps {
    connectorType: ConnectorType;
    selectedNode?: FlowNode;
}

export interface ConnectorConfigProps extends BaseConnectorProps {
    onSave?: (selectedCallNode: FlowNode) => void;
    onCreateNew?: () => void;
}

export interface ConnectorSearchConfig {
    query: string;
    searchKind: SearchKind;
}

export interface ConnectorSelectorProps extends BaseConnectorProps {
    onSelect?: (id: string, metadata?: any) => void;
}

export interface ConnectorCreatorProps extends BaseConnectorProps {
    nodeFormTemplate?: FlowNode;
    onSave?: (connectionNode: FlowNode) => void;
}